Move all filter option component into one folder
