+++ src/features/CourtCaseFilters/CourtCaseFilter.tsx
@@ -1,7 +1,7 @@
-import CaseStateFilterOptions from "components/CaseStateFilter/CaseStateFilterOptions"
-import ReasonFilterOptions from "components/CourtDateFilter/ReasonFilterOptions"
-import UrgencyFilterOptions from "components/CourtDateFilter/UrgencyFilterOptions"
-import LockedFilterOptions from "components/LockedFilter/LockedFilterOptions"
+import CaseStateFilterOptions from "components/FilterOptions/CaseStateFilterOptions"
+import ReasonFilterOptions from "components/FilterOptions/ReasonFilterOptions"
+import UrgencyFilterOptions from "components/FilterOptions/UrgencyFilterOptions"
+import LockedFilterOptions from "components/FilterOptions/LockedFilterOptions"
 import { LabelText } from "govuk-react"
 import { ChangeEvent, useReducer } from "react"
 import { createUseStyles } from "react-jss"
@@ -9,7 +9,7 @@ import { CaseState, Reason } from "types/CaseListQueryParams"
 import type { Filter, FilterAction } from "types/CourtCaseFilter"
 import { caseStateLabels } from "utils/caseStateFilters"
 import { anyFilterChips } from "utils/filterChips"
-import CourtDateFilterOptions from "../../components/CourtDateFilter/CourtDateFilterOptions"
+import CourtDateFilterOptions from "../../components/FilterOptions/CourtDateFilterOptions"
 import ExpandingFilters from "./ExpandingFilters"
 import FilterChipSection from "./FilterChipSection"
